fix: not shortest route with bidirectional nodes
